pyproject.toml
[project]
name = "ap-executor"
version = "0.1.0"
description = "AP Executor – execute Analytical Pattern operators step by step"
authors = [{ name = "DataGems EOSC" }]
readme = "README.md"
keywords = ["python", "analytical-patterns", "executor"]
requires-python = ">=3.14,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi[standard]>=0.128.0",
    "psycopg[binary,pool]>=3.3",
    "pydantic>=2.12.5",
    "orjson>=3.10.0",
    "python-dotenv>=1.1.0",
    "celery[redis]>=5.6.2",
    "redis>=6.4.0",
]

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "pre-commit>=2.20.0",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "testcontainers[postgres]>=4.14.0",
    "neoteroi-mkdocs>=1.2.0",
    "pymdown-extensions>=10.20",
    "mike>=2.1.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools]
py-modules = ["ap_executor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
